unify move tests
